levels_zones/requirements.txt
# Core Dependencies
PyQt6==6.5.0
python-dotenv==1.0.0

# Database
supabase==2.0.0
psycopg2-binary==2.9.9

# Market Data
polygon-api-client==1.12.6

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.0
black==23.7.0
flake8==6.1.0
